requirements.txt
# AWS SDKs
boto3>=1.34.0
opensearch-py>=2.5.0
requests>=2.31.0
requests-aws4auth>=1.2.3

# API Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Agent Orchestration & LLM
langgraph>=0.0.50
langchain>=0.1.0
langchain-aws>=0.1.0
crewai>=0.1.0
crewai-tools>=0.1.0

# Concurrency & Performance
aioredis>=2.0.1
redis>=5.0.0
slowapi>=0.1.9
tenacity>=8.2.3
asyncio-contextmanager>=1.0.0

# Utilities
python-dotenv>=1.0.1
opencv-python-headless>=4.9.0
pillow>=10.3.0
numpy>=1.26.4
httpx>=0.27.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
wikipedia>=1.4.0
feedparser>=6.0.0

# Monitoring & Observability
python-json-logger>=2.0.7
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger>=1.21.0


# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=4.1.0
httpx>=0.25.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
cryptography>=41.0.0

# Dev
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0
